+++ src/vue/forms/market-orders/CreateTradeOfferForm.vue
@@ -135,16 +135,17 @@
       </div>
     </template>
   </form>
-
-  <loader v-else message-id="create-trade-offer-form.loading-msg" />
+  <skeleton-loader-offer-form
+    v-else
+  />
 </template>
 
 <script>
 import debounce from 'lodash/debounce'
 
 import ReadonlyField from '@/vue/fields/ReadonlyField'
+import SkeletonLoaderOfferForm from './SkeletonLoaderOfferForm'
 import FeesRenderer from '@/vue/common/fees/FeesRenderer'
-import Loader from '@/vue/common/Loader'
 
 import FormMixin from '@/vue/mixins/form.mixin'
 import OfferManagerMixin from '@/vue/mixins/offer-manager.mixin'
@@ -175,8 +176,8 @@ export default {
   name: 'create-trade-offer-form',
   components: {
     ReadonlyField,
+    SkeletonLoaderOfferForm,
     FeesRenderer,
-    Loader,
   },
   mixins: [
     FormMixin,