+++ src/vue/modules/movements-history/components/movements-table.vue
@@ -39,19 +39,27 @@
       :key="movement.id"
     />
     <empty-tbody-placeholder
-      v-if="!movements.length"
+      v-if="!movements.length && isMovementsLoaded"
     />
+    <template v-for="index in itemPerSkeletonLoader">
+      <movements-skeleton-loader
+        v-if="!isMovementsLoaded && !movements.length"
+        :key="index"
+      />
+    </template>
   </table>
 </template>
 
 <script>
 import MovementsTableRow from './movements-table-row'
+import MovementsSkeletonLoader from './movements-skeleton-loader.vue'
 import EmptyTbodyPlaceholder from './movements-empty-list-placeholder.vue'
 
 export default {
   name: 'movement-list-renderer',
   components: {
     MovementsTableRow,
+    MovementsSkeletonLoader,
     EmptyTbodyPlaceholder,
   },
   props: {
@@ -59,7 +67,14 @@ export default {
       type: Array, /** {@link Movement} **/
       required: true,
     },
+    isMovementsLoaded: {
+      type: Boolean,
+      required: true,
+    },
   },
+  data: _ => ({
+    itemPerSkeletonLoader: 3,
+  }),
 }
 </script>