+++ src/vue/modules/requests/incoming-withdrawal-requests/components/requests-table.vue
@@ -27,7 +27,7 @@
         </thead>
 
         <tbody
-          v-if="requests.length"
+          v-if="requests.length && isLoaded"
         >
           <tr
             v-for="(request, index) in requests"
@@ -64,9 +64,13 @@
           </tr>
         </tbody>
         <empty-tbody-placeholder
-          v-else
+          v-else-if="!requests.length && isLoaded"
           :colspan="5"
           :message="'incoming-withdrawal-requests.no-history-desc' | globalize"
+        />
+        <skeleton-loader-table-body
+          v-else
+          :cells="5"
         />
       </table>
     </div>
@@ -76,6 +80,7 @@
 <script>
 import EmailGetter from '@/vue/common/EmailGetter'
 import RequestStateViewer from '../../shared/components/request-state-viewer'
+import SkeletonLoaderTableBody from '@/vue/common/skeleton-loader/SkeletonLoaderTableBody'
 import EmptyTbodyPlaceholder from '@/vue/common/EmptyTbodyPlaceholder'
 
 const EVENTS = {
@@ -87,11 +92,13 @@ export default {
   components: {
     EmailGetter,
     RequestStateViewer,
+    SkeletonLoaderTableBody,
     EmptyTbodyPlaceholder,
   },
 
   props: {
     requests: { type: Array, required: true },
+    isLoaded: { type: Boolean, required: true },
   },
 
   data: _ => ({